pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "discord-tldr"
version = "1.0.0"
description = 'Discord TL;DR by Wordcab'
readme = "README.md"
homepage = "https://github.com/Wordcab/discord-tldr"
repository = "https://github.com/Wordcab/discord-tldr"
authors = ["Wordcab <info@wordcab.com>"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
aiohttp = ">=3.8"
aiosqlite = ">=0.17"
"discord.py" = ">=2.1"
greenlet = "*"
python-dotenv = ">=0.21"
sqlmodel = ">=0.0.8"
wordcab = ">=0.3"
pytimeparse = ">=1.1.8"

[tool.poetry.dev-dependencies]
black = ">=21.12b0"
coverage = ">=6.2"
pytest = ">=6.2.5"
pyupgrade = ">=2.29.1"

[tool.poetry.scripts]
wordcab-discord-bot = "discord_tldr.__main__"
